+++ impl/src/main/java/org/jboss/seam/exception/control/impl/ExceptionHandlerComparator.java
@@ -20,7 +20,9 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.jboss.seam.exceptionhandling;
+package org.jboss.seam.exception.control.impl;
+
+import org.jboss.seam.exception.control.ExceptionHandler;
 
 import java.util.Comparator;
