+++ impl/src/test/java/org/jboss/seam/exception/control/test/ExceptionHandlerComparatorTest.java
@@ -20,11 +20,11 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.jboss.seam.exceptionhandling.test;
+package org.jboss.seam.exception.control.test;
 
-import org.jboss.seam.exceptionhandling.ExceptionHandler;
-import org.jboss.seam.exceptionhandling.ExceptionHandlerComparator;
-import org.jboss.seam.exceptionhandling.State;
+import org.jboss.seam.exception.control.ExceptionHandler;
+import org.jboss.seam.exception.control.State;
+import org.jboss.seam.exception.control.impl.ExceptionHandlerComparator;
 import org.junit.Test;
 
 import java.util.Arrays;
